+++ apps/dashboard/src/app/pay/layout.tsx
@@ -1,4 +1,4 @@
-import "../../global.css";
+import "@workspace/ui/global.css";
 import { Inter } from "next/font/google";
 import { ThemeProvider } from "next-themes";
 import { cn } from "@/lib/utils";
