+++ apps/dashboard/tailwind.config.ts
