+++ apps/playground-web/next.config.mjs
@@ -37,6 +37,7 @@ const securityHeaders = [
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  transpilePackages: ["@workspace/ui"],
   eslint: {
     ignoreDuringBuilds: true,
   },
@@ -163,10 +164,6 @@ const nextConfig = {
         source: "/_ph/decide",
       },
     ];
-  },
-  webpack: (config) => {
-    config.externals.push("pino-pretty", "lokijs", "encoding");
-    return config;
   },
 };
