--- apps/playground-web/postcss.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  plugins: {
-    autoprefixer: {},
-    tailwindcss: {},
-  },
-};
