+++ apps/portal/package.json
@@ -55,6 +55,7 @@
     "@types/tryghost__content-api": "^1.3.16",
     "@typescript-eslint/eslint-plugin": "7.14.1",
     "@typescript-eslint/parser": "7.14.1",
+    "@workspace/ui": "workspace:*",
     "autoprefixer": "^10.4.21",
     "eslint": "8.57.0",
     "eslint-config-biome": "1.9.4",
