+++ apps/portal/src/app/globals.css
@@ -1,123 +1,3 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer base {
-  :root {
-    /* bg - neutral  */
-    --background: 0 0% 98%;
-    --popover: 0 0% 100%;
-    --card: 0 0% 100%;
-    --secondary: 0 0% 90%;
-    --muted: 0 0% 97%;
-    --accent: 0 0% 93%;
-    --inverted: 0 0% 4%;
-
-    /* bg - colorful */
-    --primary: 221 83% 54%;
-    --destructive: 360 72% 51%;
-
-    /* Text */
-    --foreground: 0 0% 4%;
-    --card-foreground: 0 0% 4%;
-    --popover-foreground: 240 10% 4%;
-    --primary-foreground: 0 0% 100%;
-    --secondary-foreground: 0 0% 4%;
-    --muted-foreground: 0 0% 40%;
-    --accent-foreground: 0 0% 9%;
-    --destructive-foreground: 0 0% 100%;
-    --inverted-foreground: 0 0% 100%;
-    --link-foreground: 221.21deg 83.19% 53.33%;
-    --success-text: 142.09 70.56% 35.29%;
-    --warning-text: 38 92% 40%;
-    --destructive-text: 360 72% 60%;
-
-    /* Borders */
-    --border: 0 0% 85%;
-    --active-border: 0 0% 70%;
-    --input: 0 0% 85%;
-    --ring: 0 0% 80%;
-
-    /* Others */
-    --radius: 0.5rem;
-
-    /* Sidebar */
-    --sidebar-background: 0 0% 98%;
-    --sidebar-foreground: 0 0% 4%;
-    --sidebar-primary: 221 83% 54%;
-    --sidebar-primary-foreground: 0 0% 100%;
-    --sidebar-accent: 0 0% 93%;
-    --sidebar-accent-foreground: 0 0% 9%;
-    --sidebar-border: 0 0% 85%;
-    --sidebar-ring: 0 0% 80%;
-  }
-
-  .dark {
-    /* bg - neutral  */
-    --background: 0 0% 0%;
-    --card: 0 0% 3.92%;
-    --popover: 0 0% 0%;
-    --secondary: 0 0% 11%;
-    --muted: 0 0% 8%;
-    --accent: 0 0% 11%;
-    --inverted: 0 0% 100%;
-
-    /* bg - colorful */
-    --primary: 221 83% 54%;
-    --destructive: 360 72% 51%;
-
-    /* Text */
-    --foreground: 0 0% 98%;
-    --card-foreground: 0 0% 98%;
-    --popover-foreground: 0 0% 98%;
-    --primary-foreground: 0 0% 100%;
-    --secondary-foreground: 0 0% 98%;
-    --muted-foreground: 0 0% 63%;
-    --accent-foreground: 0 0% 98%;
-    --destructive-foreground: 0 0% 100%;
-    --link-foreground: 215.88 100% 65%;
-    --warning-text: 38 92% 50%;
-    --destructive-text: 360 72% 55%;
-    --success-text: 142 75% 50%;
-    --inverted-foreground: 0 0% 0%;
-
-    /* Borders */
-    --border: 0 0% 15%;
-    --active-border: 0 0% 22%;
-    --ring: 0 0% 30%;
-    --input: 0 0% 15%;
-
-    /* sidebar */
-    --sidebar-background: 0 0% 0%;
-    --sidebar-foreground: 0 0% 98%;
-    --sidebar-primary: 221 83% 54%;
-    --sidebar-primary-foreground: 0 0% 100%;
-    --sidebar-accent: 0 0% 11%;
-    --sidebar-accent-foreground: 0 0% 98%;
-    --sidebar-border: 0 0% 15%;
-    --sidebar-ring: 0 0% 30%;
-  }
-}
-
-@layer base {
-  * {
-    @apply border-border;
-  }
-  body {
-    @apply bg-background text-foreground;
-  }
-}
-
-@layer utilities {
-  .no-scrollbar::-webkit-scrollbar {
-    display: none;
-  }
-  .no-scrollbar {
-    -ms-overflow-style: none;
-    scrollbar-width: none;
-  }
-}
-
 code span {
   color: var(--code-light-color);
 }
@@ -129,43 +9,6 @@ code span {
 .shiki,
 .shiki span {
   background-color: transparent !important;
-}
-
-.dark .shiki,
-.dark .shiki span {
-  color: var(--shiki-dark) !important;
-  /* Optional, if you also want font styles */
-  font-style: var(--shiki-dark-font-style) !important;
-  font-weight: var(--shiki-dark-font-weight) !important;
-  text-decoration: var(--shiki-dark-text-decoration) !important;
-}
-
-/* Fix colors on auto-filled inputs  */
-input:-webkit-autofill,
-input:-webkit-autofill:hover,
-input:-webkit-autofill:focus,
-input:-webkit-autofill:active {
-  /* Revert text color */
-  -webkit-text-fill-color: hsl(var(--foreground)) !important;
-  color: hsl(var(--foreground)) !important;
-  caret-color: hsl(var(--foreground)) !important;
-
-  /* Revert background color */
-  transition: background-color 5000s ease-in-out 0s;
-}
-
-@layer utilities {
-  /* Hide scrollbar for Chrome, Safari and Opera */
-  .no-scrollbar::-webkit-scrollbar,
-  .no-scrollbar *::-webkit-scrollbar {
-    display: none;
-  }
-  /* Hide scrollbar for IE, Edge and Firefox */
-  .no-scrollbar,
-  .no-scrollbar * {
-    -ms-overflow-style: none; /* IE and Edge */
-    scrollbar-width: none; /* Firefox */
-  }
 }
 
 .styled-scrollbar::-webkit-scrollbar {
