+++ apps/wallet-ui/next.config.mjs
@@ -74,10 +74,6 @@ const nextConfig = {
       },
     ];
   },
-  webpack: (config) => {
-    config.externals.push("pino-pretty", "lokijs", "encoding");
-    return config;
-  },
 };
 
 export default nextConfig;
